requirements.txt
# RailBooker - Unified Requirements
# Global requirements for both backend and ai-gateway services

# Core FastAPI and Web Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI and ML Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
nltk==3.8.1
spacy==3.7.2

# Utilities and Helpers
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2
redis==5.0.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Environment and Configuration
python-dotenv==1.0.0
